Get results of completed Censys SQL query job

The Get Results endpoint allows you to retrieve results of a query after it has completed successfully.

censys_get_job_result(job_id, page = 1)

Arguments

job_id

Censys job id (from calling censys_query())

page

page number of paged results

Details

You must have both CENSYS_API_ID and CENSYS_API_SECRET present in the R environment for the functions in this package to work. It is highly suggested that you place those in ~/.Renviron at least for interactive work.

References

Censys SQL query syntax: https://censys.io/query; API doc: https://www.censys.io/api/v1/docs/search

Examples

## Not run: ------------------------------------
# q <- censys_query("SELECT p443.https.tls.cipher_suite.name, count(ip) FROM ipv4
#                    WHERE p443.https.tls.validation.browser_trusted=true
#                    GROUP BY p443.https.tls.cipher_suite.name;")
# censys_get_job_status(q$job_id)
# censys_get_job_result(q$job_id)
## ---------------------------------------------